+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/tobe/HandlerMappingRegistryTest.java
@@ -4,13 +4,11 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import samples.TestManualHandlerMapping;
 
 class HandlerMappingRegistryTest {
 
@@ -19,25 +17,6 @@ class HandlerMappingRegistryTest {
     @BeforeEach
     void setUp() {
         handlerMappingRegistry = new HandlerMappingRegistry();
-    }
-
-    @Test
-    @DisplayName("수동으로 매핑된 요청을 받으면 Controller를 반환한다.")
-    void should_return_Controller_when_get_RequestMapping_request() {
-        // given
-        HttpServletRequest request = mock(HttpServletRequest.class);
-        when(request.getRequestURI()).thenReturn("/");
-
-        TestManualHandlerMapping testManualHandlerMapping = new TestManualHandlerMapping();
-        testManualHandlerMapping.initialize();
-
-        handlerMappingRegistry.addHandlerMapping(testManualHandlerMapping);
-
-        // when
-        Object handlerMapping = handlerMappingRegistry.getHandlerMapping(request).get();
-
-        //then
-        assertThat(handlerMapping).isInstanceOf(Controller.class);
     }
 
     @Test
