--- mvc/src/test/java/samples/TestManualHandlerMapping.java
@@ -1,21 +1,0 @@
-package samples;
-
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
-import jakarta.servlet.http.HttpServletRequest;
-import java.util.HashMap;
-import java.util.Map;
-
-public class TestManualHandlerMapping implements HandlerMapping {
-    private static final Map<String, Controller> controllers = new HashMap<>();
-
-    public void initialize() {
-        controllers.put("/", new TestForwardController("/index.jsp"));
-    }
-
-    @Override
-    public Object getHandler(HttpServletRequest request) {
-        String requestURI = request.getRequestURI();
-        return controllers.get(requestURI);
-    }
-}
